requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
decorator==4.4.2
ffmpeg-python==0.2.0
filelock==3.12.0
future==0.18.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
greenlet==2.0.1
httplib2==0.22.0
idna==3.4
imageio==2.28.0
imageio-ffmpeg==0.4.8
Jinja2==3.1.2
llvmlite==0.39.1
MarkupSafe==2.1.2
more-itertools==9.1.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
openai-whisper @ git+https://github.com/openai/whisper.git@c09a7ae299c4c34c5839a76380ae407e7d785914
Pillow==9.5.0
playwright==1.30.0
proglog==0.1.10
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyee==9.0.4
pyparsing==3.0.9
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sympy==1.11.1
TikTokApi==5.2.2
tiktoken==0.3.1
torch==2.0.0
tqdm==4.65.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
